Making Types more modular and StaticAxes
* Some functions had some problems because of overly static typing
(UnitSRange needs to have step size of 1 but can't be specified as const
without assuming integer/float/etc.). This loosens up that requirement
and provides abstract interfaces so it's easier to add more types with
little new code.

* Thorough benchmarking showed 2 things: 1) creating a jump table or
using `fieldtype` were most performant for grabbing traits out of
Tuples. 2) fieldtype wasn't proving to be slower until to many high
dimensional indexings into Tuple. Benchmarks are mess and on other
hardrive, so not publshing these for now.