0.9 (#22)
The majority of the original goals accomplished by this package are now
covered by ArrayInerface.jl, meaning code here can be cut down
significantly. A lot of the static ranges are no longer necessary.
The reason there were different ones before was for static
indexing which required unit ranges. We have that in ArrayInterface
now so we can just have one static range that wraps normal parent
range. Similarly mutable ranges are being replaced by `MutableRange`.
Documentation is pretty bad right now but this is currently working as
a home for code in a transient state for other packages anyway.